Master merged

+++ BNK/plugin/bank_money_transaction/fermat-bnk-plugin-bank-money-transaction-deposit-bitdubai/src/main/java/com/bitdubai/fermat_bnk_plugin/layer/bank_money_transaction/deposit/developer/bitdubai/version_1/structure/BankTransactionImpl.java
@@ -6,6 +6,7 @@ import com.bitdubai.fermat_bnk_api.all_definition.enums.BankOperationType;
 import com.bitdubai.fermat_bnk_api.all_definition.enums.BankTransactionStatus;
 import com.bitdubai.fermat_bnk_api.all_definition.enums.TransactionType;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 /**
@@ -21,7 +22,7 @@ public class BankTransactionImpl implements BankTransaction {
     private String publicKeyWallet;
 
 
-    private float amount;
+    private BigDecimal amount;
 
     private String accountNumber;
 
@@ -37,7 +38,7 @@ public class BankTransactionImpl implements BankTransaction {
 
     private BankTransactionStatus status;
 
-    public BankTransactionImpl(UUID transactionId, String publicKeyPlugin, String publicKeyWallet, float amount, String accountNumber, FiatCurrency fiatCurrency, String memo, BankOperationType bankOperationType, TransactionType transactionType, long timestamp, BankTransactionStatus status) {
+    public BankTransactionImpl(UUID transactionId, String publicKeyPlugin, String publicKeyWallet, BigDecimal amount, String accountNumber, FiatCurrency fiatCurrency, String memo, BankOperationType bankOperationType, TransactionType transactionType, long timestamp, BankTransactionStatus status) {
         this.transactionId = transactionId;
         this.publicKeyPlugin = publicKeyPlugin;
         this.publicKeyWallet = publicKeyWallet;
@@ -73,7 +74,7 @@ public class BankTransactionImpl implements BankTransaction {
 
 
     @Override
-    public float getAmount() {
+    public BigDecimal getAmount() {
         return amount;
     }
