+++ BNK/plugin/wallet/fermat-bnk-plugin-wallet-bank-money-bitdubai/src/main/java/com/bitdubai/fermat_bnk_plugin/layer/wallet/bank_money/developer/bitdubai/version_1/structure/BankMoneyWalletBalanceImpl.java
@@ -1,14 +1,14 @@
 package com.bitdubai.fermat_bnk_plugin.layer.wallet.bank_money.developer.bitdubai.version_1.structure;
 
-import com.bitdubai.fermat_api.layer.osa_android.database_system.Database;
 import com.bitdubai.fermat_bnk_api.all_definition.enums.BalanceType;
 import com.bitdubai.fermat_bnk_api.layer.bnk_wallet.bank_money.exceptions.*;
 import com.bitdubai.fermat_bnk_api.layer.bnk_wallet.bank_money.interfaces.BankMoneyTransactionRecord;
-import com.bitdubai.fermat_bnk_api.layer.bnk_wallet.bank_money.interfaces.BankMoneyWallet;
 import com.bitdubai.fermat_bnk_api.layer.bnk_wallet.bank_money.interfaces.BankMoneyWalletBalance;
 import com.bitdubai.fermat_bnk_plugin.layer.wallet.bank_money.developer.bitdubai.version_1.database.BankMoneyWalletDao;
 import com.bitdubai.fermat_bnk_plugin.layer.wallet.bank_money.developer.bitdubai.version_1.exceptions.CantAddCreditException;
 import com.bitdubai.fermat_bnk_plugin.layer.wallet.bank_money.developer.bitdubai.version_1.exceptions.CantAddDebitException;
+
+import java.math.BigDecimal;
 
 /**
  * Created by memo on 23/11/15.
@@ -25,7 +25,7 @@ public class BankMoneyWalletBalanceImpl implements BankMoneyWalletBalance{
     }
 
     @Override
-    public double getBalance(String accountNumber) throws CantCalculateBalanceException {
+    public BigDecimal getBalance(String accountNumber) throws CantCalculateBalanceException {
         switch (balanceType){
             case AVAILABLE: return bankMoneyWalletDao.getAvailableBalance(accountNumber);
             case BOOK: return bankMoneyWalletDao.getBookBalance(accountNumber);
